Oops, x and y is int

+++ pathfinder.py
@@ -5,7 +5,7 @@ import math
 
 
 class position:
-    def __init__(self, x: float, y: float):
+    def __init__(self, x: int, y: int):
         self.x = x
         self.y = y
